Major refactor and support for adding origin tag to all CITracer spans.
