+++ tracer/src/Datadog.Trace/Ci/Tags/TestTags.cs
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
-namespace Datadog.Trace.Ci
+namespace Datadog.Trace.Ci.Tags
 {
     /// <summary>
     /// Span tags for test data model
